Valid integration tests example, removed unnecessary config
