--- VehicleEquipment/src/test/java/pl/zzpj/repository/rest/TestConfig.java
@@ -1,9 +1,0 @@
-package pl.zzpj.repository.rest;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan(basePackages = {"pl.zzpj"})
-public class TestConfig {
-}
