--- VehicleEquipment/src/test/java/pl/zzpj/repository/rest/VehicleEquipmentConfigSec.java
@@ -1,80 +1,0 @@
-package pl.zzpj.repository.rest;
-
-import static io.restassured.RestAssured.baseURI;
-import static io.restassured.RestAssured.given;
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.boot.web.servlet.context.ServletWebServerApplicationContext;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-
-@Testcontainers
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ContextConfiguration(classes = {TestConfig.class})
-public class VehicleEquipmentConfigSec {
-
-    @Container
-    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>(
-        "postgres:14"
-    ).withDatabaseName("zzpjdbTest")
-        .withUsername("zzpj")
-        .withPassword("zzpjpassword");
-    @LocalServerPort
-    private static Integer port;
-    @Autowired
-    private ServletWebServerApplicationContext webServerAppCtxt;
-
-    @DynamicPropertySource
-    static void configureProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", postgres::getJdbcUrl);
-        registry.add("spring.datasource.username", postgres::getUsername);
-        registry.add("spring.datasource.password", postgres::getPassword);
-    }
-
-    @BeforeAll
-    static void beforeAll() {
-        postgres.start();
-    }
-
-    @AfterAll
-    static void afterAll() {
-        postgres.stop();
-    }
-
-    @BeforeEach
-    void BeforeEach() {
-        RestAssured.baseURI = String.format("http://localhost:%s/", webServerAppCtxt.getWebServer().getPort());
-    }
-
-    @Test
-    void testConnection() {
-        Assertions.assertTrue(postgres.isRunning());
-        System.out.println(postgres.getMappedPort(5432));
-        System.out.println(postgres.getDatabaseName());
-        System.out.println(RestAssured.baseURI);
-    }
-
-    @Test
-    void testGet() {
-        given()
-            .header("Content-Type", "application/json")
-            .when()
-            .get(baseURI + "equipment/tires")
-            .then()
-            .statusCode(200);
-    }
-}
